requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu118
torch==2.7.1
torchvision==0.22.1

pyyaml==6.0.3
opencv-python==4.12.0.88
einops==0.8.1